Refactor components out of PlayerBar
