--- GarageBard/Components/PlayButton.swift
@@ -1,22 +1,0 @@
-//
-//  PlayButton.swift
-//  GarageBard
-//
-//  Created by Mitchel Cabuloy on 3/6/22.
-//
-
-import Foundation
-import SwiftUI
-
-struct PlayButton: View {
-    var isPlaying: Bool
-    var action: () -> Void
-
-    var body: some View {
-        Button(action: action) {
-            Image(systemName: isPlaying ? "pause.circle.fill" : "play.circle.fill")
-                .font(.system(size: 48.0))
-        }
-        .buttonStyle(PlainButtonStyle())
-    }
-}
